requirements.txt
pyrogram
tgcrypto
aria2p
python-dotenv
pytz
pymongo
uvloop
yt-dlp
git+https://github.com/Oxhellfire/pymegasdkrest